Copilot monitoring code.
This commit adds a Copilot monitor that checks that the time between when a
light switch is turned on and when the corresponding light turns on is within
certain bounds.

Auxiliary definitions and functions are placed in an Auxiliary.hs file to keep
the Main module clean. We can later propose adding those, or similar functions,
to Copilot proper.

Signed-off-by: Ivan Perez <ivan.perezdominguez@nasa.gov>