Fix two timing issues.
The main driver should multiply the time by 1000, since the internal monitor
counts in ms, not seconds.

The internal monitor should also consider the case where the light just turned
on or off, to handle immediate events. In a continuous-time scenario, this
should not be needed, but it is necessary here because continuous-time sampling
is driven by events, which are sparse.

Signed-off-by: Ivan Perez <ivan.perezdominguez@nasa.gov>