Implement ability to test Copilot monitors.
This commit adds a new target to the Makefile, `test`, that runs tests on the
Copilot monitors. Specifically, it finds every file under `tests/` and runs the
monitor against them. The expectation of whether the monitors should report an
error or not on each input file is made based on the filename: files
ending in `good` should pass, and files ending in `bad` should not.

The C code for the main driver for the monitors is modified to accept an
argument, which is used in place of the fixed LOG_PATH whenever available.

Signed-off-by: Ivan Perez <ivan.perezdominguez@nasa.gov>